General Updates
- Totem interaction fixed
- Added Pet interaction to item definition
- Update on roller events and roller interaction with items
- Update on public room static items
- Added catalog pets rotation
- Update voucher redeem
- Update some commands
- Added command buytokens
- Infobus hububar is now interactive

+++ Server/Game/Misc/Vouchers/VoucherManager.cs
@@ -9,6 +9,7 @@ using Snowlight.Communication.Outgoing;
 using Snowlight.Game.Items;
 using Snowlight.Specialized;
 using Snowlight.Util;
+using System.Web.UI.WebControls;
 
 namespace Snowlight.Game.Misc
 {
@@ -16,13 +17,43 @@ namespace Snowlight.Game.Misc
     {
         private static object mSyncRoot = new object();
 
+        private static Dictionary<string, List<uint>> mVouchersReedem = new Dictionary<string, List<uint>>();
+
+        private static readonly List<char> mNotAllowedChars = new List<char>() { 'i', 'l', 'o', 'w', 'I', 'L', 'O', 'W', ' ' } ;
+
+        public static void Initialize(SqlDatabaseClient MySqlClient)
+        {
+            DataTable VoucherTable = MySqlClient.ExecuteQueryTable("SELECT * FROM vouchers WHERE enabled = '1' AND uses > 0");
+            foreach (DataRow Row in VoucherTable.Rows)
+            {
+                string Code = Row["code"].ToString();
+
+                if(!mVouchersReedem.ContainsKey(Code))
+                {
+                    mVouchersReedem.Add(Code, new List<uint>());
+                }
+
+                string UserIdList = Row["user_id_list"].ToString();
+
+                foreach (string Uid in UserIdList.Split('|'))
+                {
+                    if(Uid == string.Empty)
+                    {
+                        continue;
+                    }
+
+                    mVouchersReedem[Code].Add(uint.Parse(Uid));
+                }
+            }
+        }
+
         public static bool TryRedeemVoucher(SqlDatabaseClient MySqlClient, Session Session, string Code)
         {
             lock (mSyncRoot)
             {
                 VoucherValueData ValueData = GetVoucherValue(Code);
 
-                if (ValueData == null)
+                if (ValueData == null || CheckUserOnUsedList(Code, Session.CharacterId) || !ValueData.CanReedemInCatalog)
                 {
                     return false;
                 }
@@ -76,7 +107,7 @@ namespace Snowlight.Game.Misc
                     }
                 }
 
-                MarkVoucherUsed(Code);
+                MarkVoucherUsed(Code, Session.CharacterId);
                 return true;
             }
         }
@@ -86,7 +117,7 @@ namespace Snowlight.Game.Misc
             using (SqlDatabaseClient MySqlClient = SqlDatabaseManager.GetClient())
             {
                 MySqlClient.SetParameter("code", Code);
-                DataRow Row = MySqlClient.ExecuteQueryRow("SELECT value_credits,value_activity_points,seasonal_currency,value_furni FROM vouchers WHERE code = @code AND enabled = '1' AND uses > 0 LIMIT 1");
+                DataRow Row = MySqlClient.ExecuteQueryRow("SELECT value_credits,value_activity_points,seasonal_currency,value_furni,can_reedem_in_catalog FROM vouchers WHERE code = @code AND enabled = '1' AND uses > 0 LIMIT 1");
 
                 if (Row == null)
                 {
@@ -97,8 +128,7 @@ namespace Snowlight.Game.Misc
 
                 foreach (string FurniValueBit in Row["value_furni"].ToString().Split(','))
                 {
-                    uint NewValue = 0;
-                    uint.TryParse(FurniValueBit, out NewValue);
+                    uint.TryParse(FurniValueBit, out uint NewValue);
 
                     if (NewValue > 0)
                     {
@@ -106,17 +136,51 @@ namespace Snowlight.Game.Misc
                     }
                 }
 
-                return new VoucherValueData((int)Row["value_credits"], (int)Row["value_activity_points"],  SeasonalCurrency.FromStringToEnum(Row["seasonal_currency"].ToString()), FurniValue);
+                return new VoucherValueData((int)Row["value_credits"], (int)Row["value_activity_points"],
+                    SeasonalCurrency.FromStringToEnum(Row["seasonal_currency"].ToString()), FurniValue,
+                    ((string)Row["can_reedem_in_catalog"] == "1"));
             }
         }
 
-        public static void MarkVoucherUsed(string Code)
+        public static void MarkVoucherUsed(string Code, uint UserId)
         {
+            if (!mVouchersReedem.ContainsKey(Code))
+            {
+                mVouchersReedem.Add(Code, new List<uint>());
+            }
+
+            mVouchersReedem[Code].Add(UserId);
+
             using (SqlDatabaseClient MySqlClient = SqlDatabaseManager.GetClient())
             {
                 MySqlClient.SetParameter("code", Code);
-                MySqlClient.ExecuteNonQuery("UPDATE vouchers SET uses = uses - 1 WHERE code = @code LIMIT 1; DELETE FROM vouchers WHERE uses < 1;");
+                MySqlClient.SetParameter("userlist", string.Join("|", mVouchersReedem[Code]));
+                MySqlClient.ExecuteNonQuery("UPDATE vouchers SET uses = uses - 1, user_id_list = @userlist WHERE code = @code LIMIT 1; DELETE FROM vouchers WHERE uses < 1;");
             }
+        }
+
+        public static bool CheckUserOnUsedList(string Code, uint UserId)
+        {
+            return mVouchersReedem.ContainsKey(Code) && mVouchersReedem[Code].Contains(UserId);
+        }
+
+        public static RedeemError ErrorChecker(string Code)
+        {
+            VoucherValueData Data = GetVoucherValue(Code);
+            if (Data != null && !Data.CanReedemInCatalog)
+            {
+                return RedeemError.ReedemHabboWeb;
+            }
+
+            for (int i = 0; i < Code.Length; i++)
+            {
+                if (mNotAllowedChars.Contains(Code[i]))
+                {
+                    return RedeemError.TechnicalError;
+                }
+            }
+
+            return RedeemError.InvalidCode;
         }
     }
 }
